Refactor code structure, move the primitive to the document level, add transofrm to handle instances
