+++ crates/crust-render/src/lib.rs
@@ -14,15 +14,16 @@ mod sampler;
 mod scene_cache;
 mod tracer;
 mod world;
+mod bvhnode;
 
 pub use camera::Camera;
 pub use convert::convert;
-pub use document::{DocObject, Document, ObjectList};
+pub use document::{DocObject, Document, ObjectList, Primitive};
 pub use hittable_list::HittableList;
 pub use light::{Light, LightList};
 pub use material::MaterialType;
 pub use material::*;
-pub use primitives::Primitive;
+pub use primitives::{Sphere,Triangle};
 pub use ray::Ray;
 pub use sampler::generate_cmj_2d;
 pub use tracer::{RenderSettings, Renderer};
