+++ crates/crust-render/src/world.rs
@@ -2,10 +2,10 @@ use crate::hittable_list::HittableList;
 use crate::light::LightList;
 use crate::material::Material;
 use crate::material::{CookTorrance, Dielectric, Disney, Emissive, Lambertian, Metal};
-use crate::primitives::Object;
 use std::sync::Arc;
 use utils::Color;
 use utils::Point3;
+use crate::Sphere;
 
 #[allow(dead_code)]
 pub fn random_scene() -> (HittableList, LightList) {
@@ -13,7 +13,7 @@ pub fn random_scene() -> (HittableList, LightList) {
     let mut lights = LightList::new();
 
     let ground_material = Arc::new(Lambertian::new(Color::new(0.5, 0.5, 0.5)));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         Point3::new(0.0, -1000.0, 0.0),
         1000.0,
         ground_material,
@@ -33,45 +33,45 @@ pub fn random_scene() -> (HittableList, LightList) {
                     // Diffuse
                     let albedo = Color::random() * Color::random();
                     let sphere_material = Arc::new(Lambertian::new(albedo));
-                    world.add(Box::new(Object::new_sphere(center, 0.2, sphere_material)));
+                    world.add(Box::new(Sphere::new(center, 0.2, sphere_material)));
                 } else if choose_mat < 0.8 {
                     // Cook-Torrance
                     let albedo = Color::random_range(0.5, 1.0);
                     let roughness = utils::random_range(0.0, 0.5);
                     let metallic = utils::random_range(0.0, 1.0);
                     let sphere_material = Arc::new(CookTorrance::new(albedo, roughness, metallic));
-                    world.add(Box::new(Object::new_sphere(center, 0.2, sphere_material)));
+                    world.add(Box::new(Sphere::new(center, 0.2, sphere_material)));
                 } else if choose_mat < 0.95 {
                     // Metal
                     let albedo = Color::random_range(0.5, 1.0);
                     let fuzz = utils::random_range(0.0, 0.5);
                     let sphere_material = Arc::new(Metal::new(albedo, fuzz));
-                    world.add(Box::new(Object::new_sphere(center, 0.2, sphere_material)));
+                    world.add(Box::new(Sphere::new(center, 0.2, sphere_material)));
                 } else {
                     // Glass
                     let sphere_material = Arc::new(Dielectric::new(1.5));
-                    world.add(Box::new(Object::new_sphere(center, 0.2, sphere_material)));
+                    world.add(Box::new(Sphere::new(center, 0.2, sphere_material)));
                 }
             }
         }
     }
 
     let material1 = Arc::new(Dielectric::new(1.5));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         Point3::new(0.0, 1.0, 0.0),
         1.0,
         material1,
     )));
 
     let material2 = Arc::new(Lambertian::new(Color::new(0.4, 0.2, 0.1)));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         Point3::new(-4.0, 1.0, 0.0),
         1.0,
         material2,
     )));
 
     let material3 = Arc::new(Metal::new(Color::new(0.7, 0.6, 0.5), 0.0));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         Point3::new(4.0, 1.0, 0.0),
         1.0,
         material3,
@@ -82,7 +82,7 @@ pub fn random_scene() -> (HittableList, LightList) {
         Point3::new(0.0, 7.0, 0.0),
         1.0,
     ));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         light.position(),
         light.radius(),
         light.clone(),
@@ -94,7 +94,7 @@ pub fn random_scene() -> (HittableList, LightList) {
         Point3::new(-4.0, 7.0, 0.0),
         1.0,
     ));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         light2.position(),
         light2.radius(),
         light2.clone(),
@@ -109,7 +109,7 @@ pub fn simple_scene() -> (HittableList, LightList) {
     let mut lights = LightList::new();
 
     let ground_material = Arc::new(Lambertian::new(Color::new(0.8, 0.5, 0.5)));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         Point3::new(0.0, -1000.0, 0.0),
         1000.0,
         ground_material,
@@ -126,34 +126,34 @@ pub fn simple_scene() -> (HittableList, LightList) {
                 _ => Arc::new(CookTorrance::new(Color::new(0.9, 0.9, 0.9), 0.2, 0.5)),
             };
 
-            world.add(Box::new(Object::new_sphere(center, 0.2, material)));
+            world.add(Box::new(Sphere::new(center, 0.2, material)));
         }
     }
 
     // Center spheres
     let material1 = Arc::new(Dielectric::new(1.5));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         Point3::new(0.0, 1.0, 0.0),
         1.0,
         material1,
     )));
 
     let material2 = Arc::new(Lambertian::new(Color::new(0.4, 0.2, 0.1)));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         Point3::new(-4.0, 1.0, 0.0),
         1.0,
         material2,
     )));
 
     let material3 = Arc::new(Metal::new(Color::new(0.7, 0.6, 0.5), 0.0));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         Point3::new(4.0, 1.0, 0.0),
         1.0,
         material3,
     )));
 
     let material4 = Arc::new(CookTorrance::new(Color::new(0.5, 0.5, 0.5), 0.2, 0.0));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         Point3::new(0.0, 1.0, 4.0),
         1.0,
         material4,
@@ -170,7 +170,7 @@ pub fn simple_scene() -> (HittableList, LightList) {
         0.0,
         1.0,
     ));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         Point3::new(0.0, 1.0, -4.0),
         1.0,
         material5,
@@ -182,7 +182,7 @@ pub fn simple_scene() -> (HittableList, LightList) {
         Point3::new(0.0, 7.0, 0.0),
         1.0,
     ));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         light1.position(),
         light1.radius(),
         light1.clone(),
@@ -194,7 +194,7 @@ pub fn simple_scene() -> (HittableList, LightList) {
         Point3::new(-4.0, 7.0, 0.0),
         1.0,
     ));
-    world.add(Box::new(Object::new_sphere(
+    world.add(Box::new(Sphere::new(
         light2.position(),
         light2.radius(),
         light2.clone(),
