+++ samples/teapot.ron
@@ -1,22 +1,22 @@
 (
     camera: (
         origin: (
-            e: (13.0, 2.0, 3.0),
+            e: (15.0, 3.0, 3.0),
         ),
         lower_left_corner: (
-            e: (2.9136019, -1.2262843, 3.8894577),
+            e: (4.8863316, -0.044797063, 4.174048),
         ),
         horizontal: (
-            e: (1.4097352, 0.0, -6.1088524),
+            e: (1.2295314, 0.0, -6.147657),
         ),
         vertical: (
-            e: (-0.5094205, 3.4875712, -0.11755858),
+            e: (-0.44830504, 3.4967792, -0.08966101),
         ),
         u: (
-            e: (0.2248595, 0.0, -0.97439116),
+            e: (0.19611613, 0.0, -0.9805807),
         ),
         v: (
-            e: (-0.14445336, 0.9889499, -0.03333539),
+            e: (-0.1271232, 0.991561, -0.02542464),
         ),
         lens_radius: 0.05,
     ),
@@ -40,6 +40,9 @@
                 name: "teapot",
                 object: Obj(
                     path: "./samples/teapot.obj",
+                    transform: (
+                        data: ((1.0, 0.0, 0.0, 0.0), (0.0, 1.0, 0.0, 0.0), (0.0, 0.0, 1.0, 0.0), (0.0, 0.0, 0.0, 1.0)),
+                    ),
                 ),
                 material: Disney((
                     base_color: (
